Refactored sqlite implementation
