+++ golem-worker-executor-base/tests/key_value_storage.rs
@@ -23,7 +23,7 @@ use golem_worker_executor_base::storage::keyvalue::memory::InMemoryKeyValueStora
 use golem_worker_executor_base::storage::keyvalue::redis::RedisKeyValueStorage;
 use golem_worker_executor_base::storage::keyvalue::sqlite::SqliteKeyValueStorage;
 use golem_worker_executor_base::storage::keyvalue::{KeyValueStorage, KeyValueStorageNamespace};
-use golem_worker_executor_base::storage::sqlite_types::SqlitePool;
+use golem_worker_executor_base::storage::sqlite::SqlitePool;
 use sqlx::sqlite::SqlitePoolOptions;
 use std::sync::Arc;
 use test_r::inherit_test_dep;
@@ -114,7 +114,7 @@ pub(crate) async fn sqlite_storage(
         .await
         .expect("Cannot connect to sqlite db");
 
-    let kvs = SqliteKeyValueStorage::new(pool);
+    let kvs = SqliteKeyValueStorage::new(pool).await.unwrap();
     SqliteKeyValueStorageWrapper { kvs }
 }
